+++ src/main/java/withermite/exercise_tracker_api/user_workout/WorkoutStatistic.java
@@ -12,5 +12,5 @@ public class WorkoutStatistic {
 
     @Min(0)
     @Max(20)
-    public Float subjectiveEffortValue;
+    public Double subjectiveEffortValue;
 }
